+++ README.md
@@ -564,7 +564,7 @@ see LICENSE file for details.
 
 ## Original Author
 
-**0x4m4** - [www.0x4m4.com](https://www.0x4m4.com) | [HexStrike](https://www.hexstrike.com)
+**[0x4m4](https://github.com/0x4m4)** - [HexStrike AI](https://github.com/0x4m4/hexstrike-ai)
 
 <div align="center">
